feat: move rdm_encode functions into c file
This enables inclusion in multiple other files without linker errors
